Fix Linux injection crash on newer kernels
This worked by accident, as ptrace expects `orig_eax` / `orig_rax` to be
set to -1 if we resume the inferior without resuming the current syscall,
in case there is one ongoing (which is usually the case). So this means
the instruction pointer would be subtracted by 2 in such cases, and we
could only hope for `mmap` to have valid and harmless instructions in
the 2 bytes in front of it.

Fixes #4