Improve compatibility with emulators
If we're not allowed to save the policy back to the kernel, SELinux is
enforcing, and we can set it to non-enforcing, we use that opportunity
to make the `shell` type permissive and flip the kernel back to enforcing.
The `shell` type is normally made permissive when rooting a real Android
device, so this is just a quirk needed for emulators.

Also allow zygote->shell to `sigchld`.