Use gum_script_unignore_later()
This allows us to avoid a deadlock situation where the GThread cleanup
acquires the slice allocator lock, which triggers a callback into the
runtime, waiting for the V8 lock. Meanwhile another thread is holding
the V8 lock, but a C++ callback tries to allocate memory and tries to
acquire the slice allocator lock.