Fix deadlock when pending calls exist at detach time
This was triggered by for example requesting to destroy a script and then
immediately also requesting to detach from the process.

We fix this by using a connection filter, allowing us to elegantly keep
track of pending incoming calls without having to add complexity to all
of our yielding server-side methods.