Improve injector so it injects psuedo-signed code
This means we avoid breaking the dynamic code-signing status of the target
process. It's weird we got away with this most of the time, although it
seems XNU eventually catches us, causing the target process to lose its
`CS_VALID` status. At least that's the current theory, this will need some
testing to say for sure.