Start adding support for multiple agent sessions
This will address a long-standing design flaw where each attach() request
would return the same object, or more precisely, the same shared session.
This meant that multiple clients would need to coordinate so that two
clients interested in attaching to the same process won't step on each
other in case one decides to detach() once it is done.