Hold on to Exceptor and batch early interception
This avoids lazily creating the Exceptor when the first script is
created, so we minimize our runtime impact and also improve performance.

As an added bonus we get to batch all of our setup and teardown
interception, which means improved performance and potentially lower
memory usage (depending on the platform).