Fix long-standing stability issue on i/macOS
Our Mach port allocation strategy was triggering an interesting bug in
the XNU kernel.

Mach allows a Mach port right owned by task A to be sent in a Mach
message to task B, which is typically used if task A has a receive right
and wants to give task B a send right, so it can receive messages from
it.

This is however not always feasible, e.g. in Frida's case, where there
isn't necessarily an established channel between task A and B, over
which a port right can be sent.

However, if task A and task B's task ports are available to an entity,
like frida-helper in our case, it is possible to directly insert a port
right into those tasks. The only challenge is that the API to do so,
`mach_port_insert_right()`, requires you to pick a unique port name. You
may go ahead and guess, and the API will tell you if the name is already
taken.

The port name is just a 32-bit unsigned integer, and is local to the
task, or more precisely, the IPC space. So this part was easy enough.
We had just allocated a port in that other task, which is really easy to
do through `mach_port_allocate()`, because a port name is chosen for you
by the kernel. Our strategy was simply to take that name incremented by
one, and try inserting that. Then if that fails, just keep incrementing
and trying. A few dozen system calls later and off we'd go, with freshly
punched Mach ports ready for task A and B to be able to communicate in
both directions.

This was around Easter 2013, when the Mach port transport was born. Up
until then we had been using TCP over loopback to communicate between
frida-core and the injected frida-agent. But I had been planning on
writing a transport per OS that would use a really low-level primitive
to avoid issues with firewalls, sandboxing, etc., and also give us a bit
more stealth.

I remember how I was planning to spend my Easter holiday just doing
this, and the goal was to do the pipe transport for Windows, i/macOS,
and Linux. Needless to say there wasn't much time to write these three
different backends, so I missed one important detail when writing the
Mach backend. Mach port names have a particular encoding where the lower
8 bits are used to identify which generation of IDs this name belongs
to, and the upper 24 bits is the index. My simple increment-and-retry
algorithm would thus be scrambling through the generation part and not
just the index.

Over the last 3 years that followed, I've been noticing some really odd
crashes on i/macOS, but always assumed it was caused by a memory
corruption issue. Indeed, there were a few bugs that caused memory
corruptions, but those have since been hunted down and fixed. With those
out of the way, AddressSanitizer happy, Valgrind not finding anything,
and hours on end in front of LLDB, this last corruption was nowhere to
be found.

After testing different components in isolation and binary-searching my
way through possible error-sources, it got quite clear that the Mach
pipe transport backend was somehow causing this, or at least making the
corruptions very reproducible. I also found another stability issue
caused by the Darwin library injector, but there wasn't anything weird
about that.

Turns out that trying to insert Mach port names with the generation
part and index all over the place is not really a good idea. I suspect
that the internal reshuffling that this triggers results in ports
briefly looking like they're gone, i.e. a race condition. We avoid this
kernel bug entirely by only incrementing the index and leaving the
generation part as-is.