Add support for injecting into a suspended process
We essentially move the process initialization from spawn() to attach(),
which also means that a straight spawn() -> resume() -- without attach()
in between -- will be much faster. More importantly, though, this allows
us to instrument externally spawned processes.

This is very useful for hooking posix_spawn() in an external process,
e.g. launchd, in order to add POSIX_SPAWN_START_SUSPENDED. Such an agent
would then just notify the Frida application of the spawned PID, and it
is then able to attach() to it, apply instrumentation, and then call
resume(pid).

The goal is to stop piggy-backing on CydiaSubstrate for app launching on
iOS, and this is the first step towards that goal. This also opens up
other interesting use-cases to advanced Frida users.