Add Injector constructor for inprocess use
This is for embedders that have the required permissions to inject, and
don't want or aren't able to spawn a child process. This means reduced
footprint at the cost of build system complexity.

For now this is only implemented on i/macOS. On iOS this requires the
caller process to have the `task_for_pid-allow` entitlement.