Add support for a third interaction called `script-directory`
E.g.:

    {
      "interaction": {
        "type": "script-directory",
        "path": "/Library/Frida/Scripts",
        "watch": true
      }
    }

Then plopping a new file named "twitter.js" into that directory:

    'use strict';

    rpc.exports = {
      init: function () {
        console.log('[init]');
      },
      dispose: function () {
        console.log('[dispose]');
      }
    };

It will be loaded automatically no matter which process it is in.
Note that the above watches the directory for changes, but not its
scripts.

Then by creating a file named "twitter.config" next to it containing:

    {
      "filter": {
        "executables": ["Twitter"],
        "bundles": ["com.twitter.twitter-mac"],
        "objc_classes": ["TwitterAPI"]
      },
      "watch": true
    }

This will ensure it is only loaded if the process' executable is named
"Twitter", or its bundle ID matches the above, or it has an Objective-C
class loaded named "TwitterAPI".

We also tell it that we would like to watch this particular script for
changes.