Fix injection on Android N and above
Calling dlopen() from outside a system library will now result in it
refusing to load libraries that are outside the app. We work around this
by finding the inner implementation and telling it that the caller is
somewhere inside libc.so.

Only fully implemented and tested on arm and arm64 – x86 support is only
partially implemented.