Add `env` to SpawnOptions
So it's possible to provide just a few environment variables to overlay
on top of the ones in Frida's execution environment.

This also means we can refuse to set `envp` in cases where this isn't
supported by the underlying OS implementation, like when spawning an iOS
app. It will however always be possible to provide `env`, as this is the
lowest common denominator on the platforms currently supported.