Fix libSystem initialization when DYLD_INSERT_LIBRARIES is there
In this case, `dlopen()` takes a different code path and hits this check in `ImageLoaderMachO::doModInitFunctions()`:

```
if ( ! dyld::gProcessInfo->libSystemInitialized ) {
    const char* installPath = getInstallPath();
    if ( (installPath == NULL) || (strcmp(installPath, LIBSYSTEM_DYLIB_PATH) != 0) )
        dyld::throwf("initializer in image (%s) that does not link with libSystem.dylib\n", this->getPath());
}
```

When we call it to initialize libSystem it throws an exception when loading `libc++.dylib` which is a dependency but it’s not libSystem itself, leaving c++ not initialized and leading to potential crashes.

This solution boils down to hook dyld’s `strcmp()` and make it successful in any case when called from the context of the above code.