Injector: added ability to read script source from stdin
Scenario: i recently bought an android device which needs to do something through injector.
This device for some reason that i don't want to know doesn't allow "setenforce 0" on the stock kernel.
The enforcing selinux doesn't allow frida injector to read files from outside it's context.
An higher level solution wich would bring benefit in other situations would be providing the script source straight to the injector through other ways.
Discussed with the owner both, a base64'd input to avoid issues with \n or an input to the stdin (through outputstream).
I'm gonna use this through an android app which uses Process to run a su -c command to invoke the injector. Tested and working by providing the script to the outputstream
and later by closing it (eof).