Fix iOS early instrumentation stability regression
The dynamic memory limit adjustments introduced in fbbdf4b made the
policy softening stateful, keeping track of whether a given PID that
we've got an agent in has already been policy-softened. That means a
cache invalidation bug may break exec-transitions because we fail to
entitle/platformize the post-exec version of the process. I.e. when
xpcproxy becomes the app or service being spawned.

Turns out we did not take into account the case where the invalidation
happened after we considered the exec-transition complete. This meant
that the spawned/spawn-gated process would die pretty quick, and make
attach() fail with the following error:

    Expecting to read a single byte for receiving credentials but read
    zero bytes

We fix this issue by explicitly notifying the backend when an exec has
completed. The backend is then able to flush its dispatch queue, and
eagerly clean up any injection instances for the given PID.