Remove system loop integration from frida-gadget
As we don't know exactly which initializers have run before us, there
might be pending I/O that assumes some of main() has already been run,
making this a very dangerous feature.

So far I haven't actually encountered any use-cases where we need to
iterate the system loop early, but it is still possible to emulate this
if needed: while waiting for resume, the GLib.MainContext used by GumJS
is serviced by the thread calling the initializer, so that means user
scripts can run a modal system loop if needed.

Note that this feature was only implemented for i/macOS, so removing it
means we now get the same behavior on all OSes – which is to block on the
GumJS' GLib.MainContext while waiting for resume().