frida-gadget: listen interaction config support for staving off resume() until after script has run and eternalizing
This patch adds two knobs to the frida-gadget config for the "listen" type:

- One enables configuring when the call to resume() is made that will resume
  the program to enable pushing it until client disconnect.
- The other configures each script to be "eternalized" so that they (and the
  instrumentations made by them) stay loaded indefinitely, even after the
  client disconnects.

These are necessary for setting up rpc.exports "init"-style logic for "listen",
which is useful when the program is short-lived (e.g. `ls`). In such case, two
issues currently arise (that this patch seeks to adress):

- The process exits during (or before!) the script's load/execution
  (`"on_load": "wait"` only ensures that resume() is not called before a
  _connection_ is made.)
- Scripts themselves appear to lack a proper mechanism to call resume() on
  their own, requiring it to be triggered on client diconnect. However, this is
  a catch-22 as client disconnects trigger script unloading.

With the patch applied, these cease to be breaking issues. It's not ideal to
have to disconnect to start, but I don't see another way at the moment without
much more invasive changes. This means no `console.log()` after "init", but
`new File("...", "ab")` works just fine.

Note: Technically, custom clients can call `ClientSession::eternalize_script`,
but the standard frida-tools `frida` CLI tool does not have support for it and
the logic for "when" to enable it can be a bit complicated depending on the use
case. To get around this, I've placed the eternalize call in
`ClientSession::destroy_script` (and disabled destroying the script when the
setting is enabled). This has a major benefit in that the interactive console
interface works when the setting is enabled and all instrumentations applied
through it will persist at runtime.