Fix initial device list logic on fruity host session
usbmuxd doesn’t accept any command after
starting the listening mode. The “dummy request”
we used to flush the device list makes usbmuxd
close the connection because the command is
received in the wrong state.

To avoid this while ensuring the initial device list is
as close as possible to the actual set of connected
devices, the new logic is:

- call ListDevices
- start listening for changes in the list
- try to connect to all devices in the initial list, to
ensure they’ve not been detached in the meantime
- remove detached devices from the list