Handle control_client closed by usbmuxd
usbmuxd doesn’t accept any command after
starting the listening mode. The “dummy request”
we used to flush the device list makes usbmuxd
close the connection becuase the command is
received in the wrong state.

To avoid this issue, the new logic is:

- detect if the peer has closed the connection
- in that case reopen it, and send the “dummy
request” on a secondary connection

That should work because usbmuxd handles
all clients on a single thread, so by the time
the “dummy request” is handled, the entire list
of currently connected devices should have
been sent.