Fix Gadget crash on i/macOS when loaded early
In situations where we are initialized before Foundation.

We solve this by explicitly depending on Foundation. This isn't ideal as
it means we will be pulling it into processes that don't already depend
on it. So until we're sure there's no better way, we will keep our
implementation treating it as optional.