Rework the function address resolution on Android (#370)
Use the library base address retrieved using dladdr() to find the path
in /proc/:pid/maps instead of relying on the library path. This is to
handle cases where the Android environment is running inside a
container.

Similarly, when resolving a linker function, use the already resolved
linker base address to match the correct line in /proc/:pid/maps.