Revert "Implement a new frida-pipe strategy for i/macOS"
It's flawed as we run into sandbox issues with many targets.

However, I'm not even sure that complex strategy was needed for iOS 15
after all, as the Mach ports crossing task boundaries are not guarded.
I don't have access to a jailbroken iOS 15 system to actually test this
though, so we'll just have to wait and see if somebody files an issue.

This reverts commit 1ea7da911dc18dfda9446d4ca4e5b3579117fe66.