darwin: Always host system session locally
Like we do on the other platforms.

It's a bit wasteful to spin up frida-helper just to inject it there,
when the helper isn't needed for anything else. Also, since we use
Gum.Darwin.Mapper instead of dlopen() in our injector, that makes it
hard to debug and profile code running in the system session.