darwin: Fix remote connectivity to hardened processes
Our direct TCP connection optimization must not be attempted when the
target process is a hardened process. The frida-agent treats its end of
the socketpair() as a simple file-descriptor, so it will not be prepared
for control messages. The DBusConnection knows this, as it negotiated
its capabilities during authentication, so all we need to do is ask it.

Co-authored-by: Håvard Sørbø <havard@hsorbo.no>