darwin: Fix app listing and launching on iOS >= 16
Some Mach services that were previously running as “root” are now
running as “mobile”. This means they're not in the same launchd domain,
and lookups fail.

While we could run as “mobile” too, our system session would no longer
be running as root, and that would break scripts that assume root
access. Though we could solve that by running multiple daemons, it
would introduce a lot of complexity.

So to solve this we re-route our local process' Mach service lookups
for the service names we know that SpringBoardServices and friends are
using.

Co-authored-by: Håvard Sørbø <havard@hsorbo.no>