darwin: Make spawn aware of prewarm
Since iOS 15, Darwin has the "prewarm" feature for which most frequently
used apps are being spawned ahead of time by `dasd` and kept suspended
until the user attempts to launch them, in order to save some startup
time.

This was in the way of Frida's spawn mechanism because:

- apps launched that way won't be going through launchd when expected,
causing a "timeout was reached" error, or more complex race conditions
- Frida needs to spawn a fresh process to ensure early instrumentation

This change addresses these problems by:

- ignoring prewarm spawns from the launchd agent
- killing existing prewarmed apps before attempting to spawn them

The detection of prewarmed apps is done by checking the presence of the
`ActivePrewarm=1` environment variable (libSystem makes the same
assumption).