fruity: Handle failure to fetch dyld symbols
Which happens on newer versions of iOS when trying to attach to a fully
booted app. In such a case we find the restarted dyld, and since it's in
the dyld_shared_cache we won't find any symbols for it there.

Before we deal with that however, better make sure we handle the case
where our symbol fetcher doesn't find any symbols.