Fallback to "slow" implementation of read_memory/write_memory on old Linux kernel
We faced some issues when launching a frida-server on a x86_64 box with an old Linux Kernel.
The syscalls `process_vm_readv` and `process_vm_writev` exist but are not implemented. Therefore, using those syscalls failed and the error ENOSYS is returned.

We allow the linux backend to fallback to the slower implementation of `read_memory`/`write_memory` when it faces this issue.