linux: Fix unreliable injection when libc is absent
Such as early in the process lifetime, and in targets without an RTLD.

The range chosen by our ProcessCodeSwapScope could happen to overlap
with the program's ELF header, which we would subsequently parse and
then carry on to fail in spectacular ways.

So for the case when libc is absent, instead of attempting the full
bootstrap right away, only perform mmap() and then return. Since we
now allocate enough for both the bootstrapper and the loader, we can
eliminate the second mmap() in the case when libc is already there.