meson: Add architecture compatibility automation
For supporting targets with a different architecture, such as 32-bit
processes on a 64-bit system.

Up until now this required multiple build directories, and passing in
options to specify where the other helper and agent binaries reside.

This was however quite painful, as it required a meta build system to
automate these steps. Getting incremental builds working reliably was
also quite difficult. It also exposed the user to platform-specific
implementation details.