meson: Fix usrsctp check when it's a subproject
The current compiler.has_member() implementation's handling of the
`dependencies` keyword argument only supports external dependencies,
at least for now.

Given that the list of OSes we support where sconn_len is defined is
rather small, we might as well just mirror a subset of the logic in
usrsctp's header file.