fruity: Inject on enforced debugger mapping
Add support for injecting the gadget on iOS targets where debugger
mappings are enforced and we can't flip the memory protection back to
executable from inside the target process.
In that case, the gadget configuration will have `code_signing` set to
`required` for now, until the Interceptor will be ready for enforced
debugger mappings too.